Fix how floating tags are determined (#204)
There is no guarantee that a floating tag is set, but we can recover
from that case if we hit it.

Also remove comments about needing floating tags for Freshmaker since it
is not a necessity. Additionally, Freshmaker would not have bumped the
tag (i.e. version), just the release for the same tag if we wanted to
pin to tags.

Signed-off-by: arewm <arewm@users.noreply.github.com>